+++ src/lib/index.js
@@ -44,7 +44,7 @@ let morning = {
         white : 'wh'
     },
     isMorning : true,
-    version : '0.10.25',
+    version : '0.10.26',
     map : {}
 };
 
@@ -174,7 +174,7 @@ morning.install = function (Vue, options) {
 
     if (+version[0] < 2 ||
         +version[1] < 5 ||
-        (version[1] === 5 && +version[2] < 3)) {
+        (+version[1] === 5 && +version[2] < 3)) {
 
         throw new Error('Vue version mismatch(version must > 2.5.2), please update Vue.');